xxxx

+++ pages/threejs/UI-Elements/ToolBar/ToolBar.tsx
@@ -52,16 +52,25 @@ function ToolBar(props: {
 
   function setEditMode(editMode: "translate" | "scale" | "rotate") {
     props.setObjProps((prev: TypeObjectProps) => {
-      // console.log("xxxxxxx: " + prev.info); // TODO: info ist leer, wenn wall ausgewählt wird x axis ausblenden, man kann wände trotzdem nicht breit ziehen??
-      // if (prev.info == "wall" && editMode == "scale") {
-      //   return {
-      //     ...prev,
-      //     editMode: editMode,
-      //     showXTransform: false,
-      //     showYTransform: true,
-      //     showZTransform: true,
-      //   };
-      // }
+      console.log(prev);
+      if (prev.info == "wall" && editMode == "scale") {
+        return {
+          ...prev,
+          editMode: editMode,
+          showXTransform: false,
+          showYTransform: true,
+          showZTransform: true,
+        };
+      }
+      if (prev.info == "floor" && editMode == "scale") {
+        return {
+          ...prev,
+          editMode: editMode,
+          showXTransform: true,
+          showYTransform: false,
+          showZTransform: true,
+        };
+      }
 
       return {
         ...prev,
@@ -184,7 +193,9 @@ function ToolBar(props: {
                         return;
                       }
 
-                      props.deleteObject(props.objProps.id);
+                      let result = confirm("Das Objekt wird gelöscht...");
+
+                      if (result) props.deleteObject(props.objProps.id);
                     }}
                   >
                     <DeleteForeverIcon></DeleteForeverIcon>
